add solution for cjsToEsm.mjs
