+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,44 @@
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
+import path from 'path';
+import { createRequire } from 'node:module';
+
+import { release, version } from 'os';
+import http from 'node:http';
+
+const require = createRequire(import.meta.url);
+import './files/c.js';
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = require('./files/a.json');
+} else {
+  unknownObject = require('./files/b.json');
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+
+console.log(`Path segment separator is "${path.sep}"`);
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = http.createServer((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
